backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile backend/requirements.in
#
cachetools==4.2.1
    # via google-auth
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via uvicorn
google-api-core[grpc]==1.26.3
    # via
    #   google-cloud-core
    #   google-cloud-firestore
google-auth==1.29.0
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-core==1.6.0
    # via google-cloud-firestore
google-cloud-firestore==2.1.0
    # via -r backend/requirements.in
googleapis-common-protos==1.53.0
    # via google-api-core
grpcio==1.37.0
    # via google-api-core
h11==0.12.0
    # via uvicorn
idna==2.10
    # via requests
packaging==20.9
    # via google-api-core
proto-plus==1.18.1
    # via google-cloud-firestore
protobuf==3.15.8
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyparsing==2.4.7
    # via packaging
python-dotenv==0.17.0
    # via -r backend/requirements.in
pytz==2021.1
    # via
    #   google-api-core
    #   google-cloud-firestore
requests==2.25.1
    # via google-api-core
rsa==4.7.2
    # via google-auth
six==1.15.0
    # via
    #   google-api-core
    #   google-auth
    #   google-cloud-core
    #   grpcio
    #   protobuf
starlette==0.14.2
    # via -r backend/requirements.in
urllib3==1.26.4
    # via requests
uvicorn==0.13.4
    # via -r backend/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
